+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Simplify move scoring by removing sentinel values
 * Started tracking the bench in a '.bench' file which can be updated via 'just bench'
 * Fixed a datagen bug which caused non-decisive games to always be stored with an outcome of a loss for white
+* Switched to using 'tacticals' terminology for non-quiet moves
 
 ## [7.0]